Update OpenBeerFiller.ino
Added an initial can push to start button, for one off push and added a can push at the end of the process so the cans are pushed clear of the filling area.
Pressing the start button again won't push the cans as they are already in place. If the filler is restarted, you'll have to remove the 3 empty cans, but these can be replace to the awaiting can slide.
I feel that this avoids manoeuvring the cans out of the fill area and means you can push fresh cans into place while removing cans for capping. 
Cheers Matt (BeerMatt)